refactor: separate Calories component to Form and Result
Related to #42
